Clients are treated as other objects now
They used to be read from the database at startup, placed in a list and
always used from there afterwards. Now their info gets read from the
database everytimg it's needed and the MAC address gets coupled with
the socked obejct in a clients_dict dictionary, if the client is
actually online.