adding test page with render command
This is first attempt at rendering something from brender.
I added a testpage that can be used for various things.
A button will tell the server to launch a command, the server then
tells the client ( as it woull in the future when we enable job
dispatching).
I the subprocess doesn't work when we log the output… to be
investigated.
waiting for input on that commit