Improved manager-server communication
Upon registration of a new manager, the server assigns it a uuid (which
gets stored by both parties) that is used do modify the manager
resource on the server side.

Added a Setting model for the worker, which later could contain further
configs that are currently stored in the config.py (blender paths and
settings).

When a worker connects to the manager, the count of total_workers
connected to it is updated and that info is updated on the manager
resource on the server (which then uses it for scheduling).

Both server and manager db upgrade are required.