Blender OpenGL preview render job.
This runs the Blender OpenGL render operator for outputting quick
previews. Even though they are much faster than actual rendering, they
can still take significant time and stall a machine in the meantime,
which makes them suitable as a farm job.

The OpenGL rendering requires an OpenGL context, so Blender cannot be
run in background mode for this job!