Manager: introducing Job Types
We are introducing a new job-types entry point, that will be dedicated
to storing and providing job type commands for each operating system
(previously these settings were retrieved from general manager
settings). This functionality is combined with a new worker-job-types
table, that stores the association of each worker with the job types it
can execute. If no relation is established, then the worker will
execute any task available.

We are also introducing a fix to the Blender add-on that allows
streaming very large files to the server when submitting a new job
(this was previously causing a memory leak).