Server: frame statuses on task level
We introduce a more granular way to monitor tasks that are frame-based.
Now a manager can provide the “frames” argument during a PUT request to
tasks/:id, and it will trigger the update of frames for that task. In
order to calculate frame ranges we use the frame_range_parse and
frame_range_merge utilities.