Worker: queue task updates in a persistent sqlite database
This allows the worker to queue up task updates when Flamenco Manager
is unreachable.

Note that asyncio is used for task scheduling & execution, but HTTP connections
and subprocess communication is not (yet). As a result, a blocking POST
call will block proc.stdout.read() in AbstractSubprocessCommand, and
vice versa. This can be solved by using either multithreading or asyncio
to perform those operations.