Worker: use asyncio task scheduling to ensure activities & logs are sent.
Before this comment, logs & activities would be queued, and this queue was
pushed to master when a log or activity was registered and certain criteria
were met. As a result, when no more logs/activities were registered, the
queue would not be flushed. This is now solved by scheduling a push to
master, and cancelling that task if that push happens for any other reason.