Manager: re-fetch task from Server before handing it to worker
This includes checking for any server-side changes. If there are, the
queued task is updated accordingly, and the queue is re-examined.

Connection errors, redirects and Server-side internal server errors are
ignored, and the task is considered "unchanged" in those cases. Errors
in the 400-range (Not Found, Forbidden etc.) result in the task being
marked as "canceled" in the queue. The task is NOT removed from the
queue, so that the information about the task being unavailable to workers
is explicitly available.

Furthermore, this commit introduces gocheck for unit testing, which gives
us setup and teardown functions.