Explicitly ignore job status change under-construction → queued
The job compiler already takes care of the task statuses. It also does this
in the correct order, so that when the job is in a runnable state ('queued'),
its tasks are also in a runnable state. Doing it in the Jobmanager's
handle_job_status_change() function would cause a race condition where
(some of) the tasks aren't runnable (still 'under-construction') but
the job is.