Added task_type to tasks so that workers can be limited to certain tasks
See eve_settings.py for the current list of valid task types. I think it's
a good idea to keep those specified in there for now, so that we have an
authoritative list of available task types.

The special task type "unknown" is used by the Manager when it gets a
task without task_type (i.e. created before this commit). Such tasks will
only be run by workers that explicitly support the "unknown" task type.