flamenco_project_view() now always extension properties
The extension properties for Flamenco are so tiny, that fetching them only
when necessary doesn't give us any measurable speedup. However, this
conditional availability of the extension properties does produce
unnecessarily complex code further down the call chain.

The `extension_props` parameter is still there, but now only determines
whether they are passed to the decorated function as a separate argument.
They are always fetched, and are thus always available on the project
document itself.