Introduced path replacement variables.
This allows Managers to push their path replacement variables to the
Server, so that a Client can pick it up and perform the replacement.
The Client takes the path and replaces any variable's value with the
variable name, f.e. "/render/flamenco/blah.blend" could become
"{render}/blah.blend". There must always be a slash between the variable
and the remainder of the path.

Tracker: https://developer.blender.org/T51747