Task list: store current page in window.sessionStorage
This allows the page to be reloaded while keeping the same page of the
task list open.

This is done in window.sessionStorage, because the URL fragment is already
used for other stuff, and changing the fragment to something unknown has
the side effect calling window.popstate() with a null state, which we
interpret as 'open the first job'.