Include `ffmpeg_cmd: {ffmpeg}` variable in create-video tasks
This allows the Manager to determine which FFmpeg binary is used. We were
thinking of bundling FFmpeg with the worker, but it turns out that the
binaries for different platforms may or may not have high-bit video support
(it's not documented well), so we wouldn't know exactly what we'd be
bundling. This made me decide against it (at least for now) so the user
will have to install FFmpeg and configure its binary location.