Added `blender-video-chunks` job type
This job type is meant for encoding the edit of a film from the video
sequence editor. In the Blender Cloud add-on, configure it to use
Matroska video and be sure to configure an audio codec if you want audio
to be in the final file.

The job consists of the following tasks:
    - Per video chunk, render the frames as PNG and combine each chunk
      into a Matroska (MKV) video file.
    - Concatenate all MKV files into one file.
    - Render the audio, convert to AAC.
    - Mux audio and video into a single MKV file.
    - Move that file to the final location, and name it with the current
      date, a counter to make the filename unique, and the name of the
      blend file.

Requires Blender Cloud add-on 2.0 or newer, and Flamenco Worker 2.2 or
newer.