Fix T51803: allow cancellation of individual tasks
This required a little change in the task status → job status propagation.
The code assumed that a task going from `cancel_requested` → `canceled`
was part of the cancellation of an entire job. This no longer is the
case, so now the job is only cancelled if it was `cancel_requested` before.