Create an RNA override file for Blender renders
When the job settings contains a key `'rna_overrides'` it is assumed to
be a list of strings. Those strings are concatenated into a Python file,
which should be loaded after the blend file was loaded (and before actual
rendering starts).

Does not contain any way to set those settings yet, apart from mucking
around in the MongoDB directly.