Allow users to request task log files from Manager
Nowadays the Manager keeps the log files, and the Server only stores the
last few lines of logging; the entire log files can be viewed on the
Manager directly. This is fine, unless a user doesn't have access to the
Manager itself.

This commit allows users to request the Manager to send its logfile to
the Server. These requests are queued, and the next time the Manager
sends task updates it receives those requests and acts accordingly.

Uploaded log files are compressed by the Manager, and stored as-is on
the project's storage backend until the task is requeued & reclaimed by
the Manager (in which case we act as if it's gone, but the file is still
there; it's just the reference that's erased).

The task's logfile can only be requested in certain states (all except
queued, claimed-by-manager, and active). This prevents a task file from
being requested while the task is still active (and thus this feature
misused for realtime updates).