Added new 'fail-requested' job status
Instead of failing a job immediately, it can now also go to
'fail-requested' status. This should be done when there are still tasks
that are managed by the Manager (like 'active', 'claimed-by-manager', etc.)
and that will go to 'cancel-requested' rather than 'canceled'.

After a task has gone to 'canceled' status and when there are no more
'cancel-requested' tasks, the job will go from 'fail-requested' to 'failed'.