Automatically delete jobs stuck in `waiting-for-files` status
These jobs are waiting for an external PATCH call to initiate job
compilation, queueing, and execution. If this PATCH call doesn't come,
the job is stuck in this status. After a certain time of waiting, this
function will automatically delete those jobs.

Be sure to add a schedule to the Celery Beat like this:

    'remove_waiting_for_files': {
        'task': 'flamenco.celery.job_cleanup.remove_waiting_for_files',
        'schedule': 3600,  # every N seconds
    }