Allow CORS requests from a Manager to the JWT token generator URL
These requests are authenticated by the user session. In other words, the
user should visit Flamenco Server in their webbrowser and log in, before
a token is generated. While the session is valid any number of tokens can
be generated.

To avoid CSRF attacks, a request that uses the session to authenticate the
user MUST include an expiry timestamp