Support version 2 of Flamenco Manager variables
A restructuring of the Manager variables was necessary to allow
different values for users and workers. Until now both had the same
paths, which is fine when both are located in the same studio. However,
when deploying workers on a cloud infrastructure (Azure, AWS, etc.) this
becomes cumbersome.

When a Manager does not declare a `settings_version`, version 1 is
assumed. This commit introduces version 2.

The naming 'path replacement variable' was confusing people. Now each
variable has a mandatory property 'direction' that has one of the
following values:

- `"one"`: unidirectional variable. `${VARIABLENAME}` is replaced by the
  variable value for the appropriate platform. This used to be the
  'variable'.
- `"two"`: bidirection variable. As described above, but also the value
  is replaced by `${VARIABLENAME}` by the Blender Cloud add-on when
  sending a job to Flamenco Server. This used to be the 'path
  replacement variable'.

Each variable now has a setting 'audience', which can be either
'workers' or 'users' to be applied only to those, or 'all' when the
value is the same for both.