Re-queueing a task now erases its `failed_by_workers` list
The 'failed_by_workers' list is also erased when the entire job is
re-queued.

Erasing the list ensures that workers that previously failed the task
can try again. This is essential when the task was failing due to
external conditions, for example failure due to a bug in Flamenco Worker
that's since been resolved.