Added `delete_orphan_task_logs` CLI command
It fetches the ID and the Task ID for *every* task log entry, and performs
a 'does this task exist' check in Python. The logic is in Python because I
couldn't find a MongoDB query that allows "foreign key" lookups and delete
documents where this cannot find the referenced document.